requirements.txt
# torch在pytorch官网中下载cuda版本
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision
hydra-core
lightning[extra]
numpy

